Remove casting feature from `maybePrime`.
`maybePrime` used to implicitly cast between any Integral types. This is
wrong as explained here: codereview.stackexchange.com/a/176214/149320

In short, a user may exploit the fact that a type cast would modify
value if the value does not "fit" in the resultant type. For example,
Word8 would only fit primes smaller than 255, so one can give a larger
prime and have it trimmed down.